Refactor installer and add tests
