+++ src/main/kotlin/host/flux/cli/install/Installer.kt
@@ -1,0 +1,5 @@
+package host.flux.cli.install
+
+interface Installer {
+    fun installLatest(): String
+}
